don't sync 11.0 to ODP mirror

+++ include/nightly.sh
@@ -84,7 +84,7 @@ build()
     fi
 
     # Sync with opendesireproject.org
-    if [ "$DEVICE" = "ace" ]
+    if [ "$DEVICE" = "ace" ] && [ "$CM_VERSION" != "11" ]
     then
         rsync -avvruO -e ssh --delete --timeout=60 $LOCAL_BASE_DIR/$PROJECT_DIR mustaavalkosta@opendesireproject.org:~/downloads/$PROJECT_DIR
         ssh mustaavalkosta@opendesireproject.org 'cd ~/ota-scanner/ && python scanner.py'
